requirements.txt
appdirs==1.4.0
Django==1.10.4
django-filter==1.0.1
djangorestframework==3.5.3
Markdown==2.6.8
packaging==16.8
psycopg2==2.6.2
pyparsing==2.1.10
six==1.10.0
